requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
tzdata==2022.1
psycopg2_binary==2.9.3
redis>=3.5
celery>=5.0
requests==2.27.1
django-extensions==3.1.5
django-shell-plus==1.1.7
ipython==8.2.0
python-dateutil==2.8.2
gunicorn==20.1.0
pylint==2.14.4
boto3>=1.15.0
mysqlclient==2.1.0
psycopg2_binary==2.9.3

